# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: DS Education Platform API
  version: 0.1.0
  description: |
    Headless learner API for the DS Education Platform (CME).

    Every request carries two credentials:
    - a **bearer token**, validated server-side against the project's own
      Keycloak realm (signature, issuer, audience, expiry — ADR-0003), and
    - the **X-DS-Project header**, identifying which host surface is calling
      (ADR-0007). The project binding decides the Keycloak realm and pins the
      tenant; a token from another realm is rejected regardless of validity.

    All compliance decisions are server-side. The client renders; it is never
    the source of truth for anything affecting a CME point.
servers:
  - url: /
    description: Relative to wherever the API is deployed.
security:
  - bearerAuth: []
...
paths:
...
  /admin/participants:
    get:
      tags: [admin]
      operationId: adminListParticipantAccounts
      summary: The people who learn with this customer
      description: |
        Deliberately **not** the same list as `/admin/learners`. That one is a
        row per enrolment — a person per course — and can only ever show
        somebody who has already started something. This is a row per *person*,
        including one created two minutes ago who has enrolled in nothing,
        which is exactly when an administrator most needs to find them.

        Scoped by `user_customers` under RLS: a person with no membership in the
        caller's customer is not listed, whatever else is true of them.
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
        - name: q
          in: query
          required: false
          description: Substring of the name or e-mail address. Absent means all.
          schema: { type: string }
      responses:
        "200":
          description: The participants.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ParticipantAccount" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      tags: [admin]
      operationId: adminCreateParticipant
      summary: Create a participant, and receive their one-time password
      description: |
        The customer is taken from the authenticated session, never from the
        body — accepting it would let one edited request plant a participant
        inside another tenant.

        **The response carries the only copy of the password.** It is stored as
        an Argon2id hash and no endpoint returns it again; a caller that
        discards this response has to reset it rather than look it up. The
        account is created with `mustChangePassword` set, because a password an
        administrator read off a screen is a password an administrator knows.

        Not self-service: the caller is an authenticated member of staff.
        `CLAUDE.md` §3 defers open registration and this does not undo it.
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, firstName, lastName]
              properties:
                email: { type: string, format: email, maxLength: 320 }
                # Both required. A Teilnahmebescheinigung prints a name and
                # cannot be issued without one, so a nameless participant is
                # somebody who finishes a course and cannot be given the point.
                firstName: { type: string, minLength: 1, maxLength: 100 }
                lastName: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "201":
          description: Created. The password is shown here and nowhere else.
          content:
            application/json:
              schema:
                type: object
                required: [userId, temporaryPassword]
                properties:
                  userId: { type: string, format: uuid }
                  temporaryPassword: { type: string }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /auth/participant/sign-in:
    post:
      tags: [auth]
      operationId: participantSignIn
      summary: Sign a participant in with an e-mail and a password
      description: |
        Only for a project whose `identity_provider` is `local`. One whose
        learners are federated is refused here with the same 401 as a wrong
        password — telling a caller "this customer uses a different identity
        provider" describes their infrastructure to anybody who asks.

        On success the session arrives as an **httpOnly cookie**, never in the
        body: a token script can read is a token an XSS bug can exfiltrate.

        Every failure — wrong address, wrong password, locked account, disabled
        account, federated project, malformed body — is the same 401 with the
        same German message. Any distinction is an account-enumeration oracle,
        and the list it would build is a list of physicians.
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 320 }
                password: { type: string, maxLength: 1024 }
      responses:
        "200":
          description: Signed in. The session is in the `Set-Cookie` header.
          content:
            application/json:
              schema:
                type: object
                required: [mustChangePassword]
                properties:
                  mustChangePassword:
                    type: boolean
                    description: |
                      The participant has not yet chosen their own password —
                      an administrator created or reset it. The portal shows
                      the change screen and refuses to go further until it is
                      done. Re-derived on every `GET me`, so a reload cannot
                      skip it.
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "429":
          description: |
            Too many attempts. The tightest limit in the platform — this is the
            only unauthenticated write on the learner plane.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  /auth/participant/me:
    get:
      tags: [auth]
      operationId: getParticipantSession
      summary: Who the caller is, and whether they owe a password change
      description: |
        The portal's "am I still signed in?" on load. The session cookie is
        `httpOnly`, so a round trip is the only way to ask.

        `mustChangePassword` is re-derived from the database here rather than
        remembered from the sign-in response, and that is what makes the
        requirement a requirement: otherwise a participant shown the change
        screen could press F5 and land in the catalogue, because the session is
        already perfectly valid.

        Omitted entirely for a federated participant — there is no local
        password, and reporting `false` would imply there is one and it is fine.
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
      responses:
        "200":
          description: The caller.
          content:
            application/json:
              schema:
                type: object
                required: [userId, subject, customerId, role]
                properties:
                  userId: { type: string, format: uuid }
                  subject:
                    type: string
                    description: |
                      The identity provider's own `sub` for this session. The
                      only value that ties a session here to a user in the
                      customer's realm, which is what a support conversation
                      about a broken federated sign-in needs (P54-01).
                  email:
                    type: string
                    description: |
                      The caller's own address, when the identity provider
                      releases it. Absent rather than null when it does not —
                      a null would read as "we hold none".
                  customerId: { type: string, format: uuid }
                  role: { type: string }
                  mustChangePassword: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /auth/participant/password-reset:
    post:
      tags: [auth]
      operationId: requestParticipantPasswordReset
      summary: Ask for a password-reset link
      description: |
        **Answers 202 in every case** — an unknown address, a project whose
        participants sign in through a customer's Keycloak, a project with no
        SMTP settings, and a link that actually went out.

        That is not politeness. Asking whether a given physician has an account
        with a named pharmaceutical company is close enough to health-adjacent
        information about a named person that this form must not answer it. The
        service behaves the same way on its side: no token is minted when there
        is nowhere to send one, so there is no difference in work done either.

        The link points at the tenant's own path, built from the request's
        `Origin` when the deployment allows it and from the first configured
        origin otherwise — **never** from the body. A reset flow that lets the
        caller name the link's host delivers a real token to a real inbox
        pointing at a page the attacker controls.

        Rate-limited per IP rather than per address: a per-address limit would
        itself answer "is this address being asked about a lot".
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, maxLength: 320 }
      responses:
        "202": { description: Accepted. Says nothing about whether an account exists. }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/participant/password-reset/confirm:
    post:
      tags: [auth]
      operationId: confirmParticipantPasswordReset
      summary: Spend a reset link and set a password
      description: |
        Public, because the token **is** the credential — whoever holds it has
        by definition no session.

        The password policy is checked before the token is spent, so a rejected
        password does not consume a single-use link and strand somebody.

        One answer for a link that expired, one already spent and one that never
        existed: a distinguishable "already used" would confirm to somebody
        holding a forwarded link that it was once real. The lifetime is
        `RESET_VALID_MINUTES` (60) and it is enforced on redemption — see
        P39-01, which is the record of what happens when it is not.
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string, maxLength: 512 }
                newPassword: { type: string, maxLength: 256 }
      responses:
        "204": { description: Set. The link is spent. }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/participant/password:
    post:
      tags: [auth]
      operationId: changeParticipantPassword
      summary: A participant chooses their own password
      description: |
        Requires the **current** password even though the caller already holds a
        valid session. A session is a bearer credential: without this, a cookie
        captured on a shared clinic computer would be enough to lock a physician
        out of their own CME record.

        The policy is the platform's single one (`checkPassword`): at least 12
        code points, not containing the account's own name or address.

        Does not revoke the caller's other sessions, deliberately — this runs
        immediately after a sign-in that reported `mustChangePassword`, and
        ending that session would sign somebody out of a change they just
        completed.
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, maxLength: 256 }
                newPassword: { type: string, maxLength: 256 }
      responses:
        "204": { description: Changed. }
        # One answer for a wrong current password, a disabled account and a
        # federated one alike — none is something the client can usefully act
        # on differently.
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Keycloak access token. Verified against the realm bound to the
        X-DS-Project header — issuer and audience come from that binding,
        never from the token (ADR-0003).
  parameters:
...
    ProjectHeader:
      name: X-DS-Project
      in: header
      required: true
      description: |
        Project slug identifying the calling host surface (ADR-0007). Pins the
        tenant, and on the learner plane also resolves the Keycloak realm to
        validate the bearer token against.

        **How a bad slug is answered depends on which plane asked**, because the
        two callers know different things already (P22-01):

        - *Learner plane* (bearer token): an unknown **or unbound** slug is a
          generic `401`, never a `404` — whether a project exists is not a fact
          an anonymous caller should be able to enumerate, and a project with no
          Keycloak binding cannot authenticate anybody in any case.
        - *Staff plane* (session cookie, ADR-0012): an unknown slug is a `404`
          carrying `detail`. The caller is already authenticated and the
          platform knows who they are, so naming what was not found is both
          honest and safe. A staff session needs no identity provider at all, so
          a project **without** a Keycloak binding resolves normally here —
          answering 401 for that locked operators out of every tenant-scoped
          console screen on a project the console itself had just created.
        - Either plane, **header absent**: `422` with `detail`. The header is
          required; omitting it is a malformed request, not a failed
          authentication, and answering 401 makes a console send the operator
          back to a login form they never left.

        A caller who is authenticated but holds no grant reaching the resolved
        customer gets `403` on both planes.
      schema: { type: string }
  responses:
...
    Forbidden:
      description: Authenticated, but no grant reaches this tenant.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...